Rebuild docker image when Dockerfile is changed
freshmaker is able to build a docker image when changed Dockerfile is
detected. freshmaker listens on fedmsg bus and handles message from
topic git.receive. When Dockerfile is changed, freshmaker will start to
build docker image in Koji immediately. Note that, freshmaker does not
wait for the task to finish, it continues to handle next coming message
instead.

Signed-off-by: Chenxiong Qi <cqi@redhat.com>