Add method for querying images with specific srpm_name
This commit adds method which given set of repositories and a srpm_name
returns image objects from those repositories which have rpms coming
from source rpm with given name. This makes use of rpm_manifest entities
in containerImages.

By default unpublished or deprecated images are ignored. We also limit
the results to images with latest tag to ignore older images. In the
future we should improve this to include any floating tags not just
latest.