Re-initialize database
The migration of 85906960545_event_search_key_and_type.py doesn't
provide an approach for altering existing records to add new columns
which don't allow NULL value, thus db upgrade fails when we do 'db
upgrade', this blocks us to generate any future migration.

Since there is no Freshmaker instance is deployed by now, we have no
real data to be maintained, and with the existing schema we can't give
reasonable 'search_key' and 'event_type', re-initialize the database is
an easy solution to workaround this (compare to changing 'search_key'
and 'event_type' to be nullable).