Add handler for brew build container task state change event
In BrewSignRpmHandler, we record all the planned Docker image builds
into database. They have the state PLANNED and they have the
ArtifactBuild.dep_on set properly according to dependencies between
these images.

When the build container task state is changed in brew, we should update
the build state in db, and if the task closed successfully, we should
rebuild containers that depends on it. At this moment, we only log the
builds which need to be rebuilt.