Auto acquire Kerberos credential
Two ways to acquire credential automatically. For running freshmaker in
a server, client keytab is used to acquire and store ticket in a
specified ccache. For running in local development, especially for
making a demo, kinit with personal user principal is required. In both
of these cases, ccache can be re-initialized again with new ticket when
old one expires.

Signed-off-by: Chenxiong Qi <cqi@redhat.com>