Start image rebuild when compose is done
Event parser and handler are created for handling ODCS state change
event. Waiting for and handling DONE event, that is compose for
rebuilding docker images is generated.

Original looping, that is used for blocking and waiting for generating
compose, is removed. In the meanwhile, repository file, that can be
passed to OSBS and used to build images, is returned as part of a new
compose created in ODCS. So, URL pointing to that repository is stored
in artifact build's build_args for saving time to retrieve compose
again when start to rebuild.

Signed-off-by: Chenxiong Qi <cqi@redhat.com>