Generate the .repo file with Pulp repositories which are needed for Container image to be rebuild.
The Lightblue code now sets 'content_sets' list for every image we are going to rebuild. This
list contains all the content-sets the image needs to be available during the build.

While recording the images, we ask ODCS to generate the .repo file with these content_sets.

Later we use that .repo file as extra repo file for container rebuild.