Send message when rebuilt images are found
Message will be sent to topic images.found once all images that are
needed to be rebuilt are found (before starting to rebuild).

Message body contains complete event and associated builds information,
each of the builds is mapped to each image to be rebuilt.

json method of Event and ArtifactBuild are modified to contain event and
build URL.

Signed-off-by: Chenxiong Qi <cqi@redhat.com>