Tweak lightblue queries to account for schema change
rpm manifest is being moved to a new entity. Fortunately containerImage
contains reference in "rpm_manifest" top level field. So we can just use
it instead of parsed_data.rpm_manifest in the same way. The reference is
an array though and should have one object with "rpms" key which is
where our list of rpm objects actually is.

This change tweaks the queries to use the new reference field instead of
parsed data.