Handle clone_distgit_repo exception.
This exception can be handled in case the dist-git commit from which the container
image has been built does not exist in dist-git anymore. This is probably an issue
in the workflow or dist-git itself, but Freshmaker should handle that.

Before this commit, Freshmaker marked whole Event as failed.

After this commit, it will only mark particular container image build as failed, but
it will still rebuild the other images.