Do not limit the LB query for `leaf_container_images` to published images only.
Using the `leaf_container_images`, the user of LB class defines the list of NVRs
of container images which should be found in LB and rebuilt later. Therefore it
does not make sense to limit such list and return only published images, because
it is the caller responsibility to find out if the image is published or not
and filter it out if he needs it.