Move ODCS code from ErrataAdvisoryRPPMsSignedhandler to FreshmakerODCSClient.
Currently, half of the ErrataAdvisoryRPMsSignedHandler is code to ask for various
ODCS composes. This is wrong from following reasons:

- The ODCS related code is not special for this particular handler and might be
  used in the future by other handlers.
- It makes ErrataAdvisoryRPMsSignedHandler longer than needed and therefore
  harder to read/understand.
- It makes the tests for ErrataAdvisoryRPMsSignedHandler longer and harder
  to understand.

This commit also removes the "boot.iso" compose type from Freshmaker, because we
have workarounded the boot.iso generation and we have no plans to support boot.iso
generation in ODCS ever because of its complexity.

There should be follow-up of this commit to further change the new FreshmakerODCSClient
API to name the methods better and make them even more general, but for simplicity,
it is not done in this commit - this commit only moves the code around without
changing it.