Publish event.state.changed.min message
Some clients load incoming UMB messages into an environment variable.
Since environment variables have a system dependent size restriction,
the existing event.state.changed message can easily exceed it.

This change makes Freshmaker publish an **additional** message that is
much smaller. It does so by omitting the list of builds associated with
the event. Instead, it provides a summary of total builds count as well
as builds count for each state. Clients loading received messages in
environment variables should use the new message.

Although the new message provides sufficient details for clients to find
out the list of builds, some clients already exist that rely on this
information to be in the message directly. For this reason, both
messages are now sent.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>