Fetch just container.yaml and content_sets.yml instead of cloning whole dist-git repo.
Currently, Freshmaker is cloning the git repository of every image it is going
to rebuild to find out the content of container.yaml and content_sets.yml to
decide whether it needs to generate Pulp composes itself or if the OSBS can
do it iself.

Cloning whole repository is slow and downloads unnecessary files.

In this commit, new `get_distgit_files` method is introduce which uses
"git archive" instead of "git clone" to get only particular files which
are needed by Freshmaker from the dist-git repository.

This is much faster and also does not consume any extra resources.