Do not rebuild container images with RPM which is already at least at the same version as the one in advisory.
If all the SRPM NVRs in the container image are newer than matching
input NVRs, the container image is filtered out from the `images`
list.

For example: The httpd-2.4-1 RPM is released together with
httpd-container. In this case, Freshmaker would try to rebuild
httpd-container, because it contains httpd package. But this is not
needed, because latest httpd-container already contains that updated
package. Therefore we filter it out in this method.