Refactor "handlers" layout.
This commit changes the directory structure for Freshmaker handlers
as well as their names. The reason is that the current structure
presumes there is just single handler for each Freshmaker Event,
but this will not be true for new features we are working on.

It is also not clear from the handler's name what is the goal of
the handler which makes the code harder to understand.

In this commit, we change the layout and names like this:

  - Handlers are now grouped according to the main service which
    they use when handling the Event - for example handlers
    resulting in Koji builds use 'koji' directory. Handlers updating
    just the Freshmaker's internal DB uses 'internal' directory, ...
  - Handlers are now named like DoSomethingOnSomeEvent. This makes
    it clear what is the goal of a handler.

For more information, you can read:

https://docs.google.com/document/d/1dx6cuFblaZlf0BZtPU6kv-ZzNJgf2JHWQ-VfSDuHVyM/edit#