Fix RebuildImagesOnODCSComposeDone to check all the images and not just those in the first batch.
The RebuildImagesOnODCSComposeDone currently checks only the images in the first
building batch (images with `dep_on == None`). This is wrong, because there might
be a situation when even image with some parent image (dep_on set to another image)
waits for a ODCS compose.

The chance for this to happen are small, because usually the ODCS compose needed
by child images are done by the time the matching parent image is built, but with
the change of the code to not wait for result of Pulp compose, we might never know
for sure.