Generate Pulp repos in case we are generating extra ODCS compose for unpublished image.
In case when image is unpublished, we need to generate an extra ODCS compose with
possibly unpublished RPMs which exist only in Koji so far.

We can pass this extra compose to Koji using "compose_ids" build option.

But the issue is that when "compose_ids" is set, Koji/OSBS completely ignores the
Pulp repos defined in the content_sets.yml file and it overrides it by whathever
we passes to it using compose_ids.

Because of that, we need to generate also the Pulp repos on Freshmaker's side
in this case and pass it together with extra compose in the compose_ids.