Include full RPM manifest in unpublished images, so we can generate extra ODCS compose with them.
Usually we do not store complete RPM manifest, but when
image is unpublished, we need complete RPM manifest in order
to check for possible unpublished RPMs.

We do not want to get the complete manifest for every container
image, because it is relatively big, so fetch it only when needed.