Use ArtifactBuild.dep_on to find out the rebuilt_nvr of the parent image.
Currently, we use `build_args["parent"]` to find out the rebuilt_nvr of
parent image. This has an issue in case we want to change the rebuilt_nvr
in the middle of a rebuild. In such case, we would need to change it
in two places - in the parent image and also in all its children images.

With this commit, we just take rebuilt_nvr only from the parent image
and do not store it also in children images. Therefore the possible
change of rebuilt_nvr is much easier.

The `build_args["parent"]` has been renamed to "original_parent"
to better express its meaning.