Do not overwrite "global" whitelist with handler-specific whitelist in handler.allow_build().
The `conf.handler_build_whitelist.get()` returns directly the dict from the whitelist.
We later called .update() on that dict which effectively changed it in
the conf.handler_build_whitelist.

In this commit, we created copy of that dict instead so the update() does not get
it into the Config class itself and stays local in allow_build method.