Store Bob builds as new ArtifactBuildType IMAGE_REPOSITORY and handle Bob response.
The artifacts we pass to Bob are not really particular container images,
but they are the names of the container image repositories. Bob then finds
out the latest image in the repository and rebuilds it.

Therefore we should have new ArtifactBuildType for them - IMAGE_REPOSITORY.

This will also fix issues with Botas handling Bob events, because Botas would
think the IMAGE artifacts are real images and try to create advisories for them.

In this commit, Freshmaker also handles the Bob response with list of 'impacted'
container repositories and store the repositories into the database, so they
can appear in the API and in the future in Estuary.