Add Event.requested_rebuilds to store the artifacts requested to rebuild for manual rebuild.
Freshmaker user can request manual rebuild of some artifacts (currently
just container images). We did not store the list of artifacts Freshmaker
requested anywhere which makes it hard to debug possible issues and also
forces requester to remember what were the container images he asked for.

In this commit, new `requested_rebuilds` column is created to store
that list and it is also provided in REST API.

The name of this column is artifact-type agnostic so we can use the same
field also for other artifact types (like image repository, modules, ...).