Add new "/verify-image" and "/verify-image-repository" API endpoints.
These are used to verify whether all conditions are satisfied for particular
container image repository or container image. This will allow container
image maintainers to check if their images are set up properly to be considered
for a rebuild by Freshmaker.

This saves me lot of time when debugging these issues with maintainers.