Do not use Lightblue "deprecated" field. Use "release_categories" instead.
The "deprecated" field in Lightblue is not used anymore and it is not
filled correctly by COMET. Instead, there is new "Deprecated"
release_category. See METAXOR-2215 for more info.

We have never actually used "deprecated" option anywhere in Freshmaker
code, because we always used the default False value to query only
for non-deprecated images.

In this commit, the "deprecated" argument is removed from Lightblue
class completely. If we ever need to query for deprecated images,
we need to use "Deprecated" release category.