Retry the `Errata._errata_authorized_get()` call.
For so far unknown reasons, Errata Tool sometimes replies with
HTTP 200 OK to authorization request (the body of that response
is not known right now). This leads to `MutualAuthenticationError`
exception in `requests_kerberos`, because it is not expected
response in this phase of Kerberos auth.

As simple workaround, it should be enough to catch that exception
and try retrying the Errata query again. This is what this commit
does.