Do not mark Beta or Tech Preview images as "latest_released".
In case when images with the same name-version but different release
appears in the Lightblue result, the deduplication code will choose
the latest_released image and replaces the images with lower Release
with the one which is marked as latest_released.

For example if there is "app" container with "foo-1-1" parent,
but latest release of foo is "foo-1-2", the "app" container will be
rebuilt against "foo-1-2". This is here, because we cannot release
multiple versions of "foo", so we need to use just single one.

Before this commit, even Beta or Tech Preview images were marked
as "latest_released". That means the "app" could be upgraded
from Generally Available image to Beta images which is wrong.

In this commit, we do not mark Beta and Tech Preview images
as "latest_released" which prevents the deduplication code
to use them as a replacement for "app" parent image in our
example.