Do not mark the event failed in-case of a service failure
Currently, when Freshmaker fails because of some DNS error while handling a build, it marks the entire event as failed. That is, all successfull builds are also marked as failed. This change changes this behavior by only marking the current build in consideration as failed and not the entire event.