Fix the switched `depending_events` and `depends_on_events` lists.
It seems to me that those REST API fields are switched. If Event A
depends on B, the B is listed in A's `depending_events`, but it
should be listed in the `depends_on_events`.

This commit fixes this, adds more comments to the code and also
renames `depends_on_events` to `depending_events`.