Remove unused and unmaintained parsers and handlers.
We originally thought Freshmaker will be used to handle modules
and container in the Fedora, but we have never reached that goal.

All this code has been written 3 years ago and nobody touched it
since that. It also most likely does not work at all, because modularity
changed dramatically since that. I'm therefore removing this old code:

- MBS and PDC classes.
- All the modules related handlers and parser.
- All the Git related handlers and parsers.
- The pdc_* and mbs_* configuration fields.