Return both Koji build NVRs and SRPM NVRs in the `Errata.get_builds()`.
This fixes an issue when some SRPMs are not shipped by policy. In this
case, those SRPMs would be missing from the `get_builds()` list.

This commit simply combines Koji build NVRs with SRPM NVRs. For normal
RPMs, these two values will match. For modules, the SRPM NVRs will
contain the list of SRPMs included in a module.

As a side-effect, the `get_builds()` will now contain also the list
of module build NVRs, but this is not wrong, since module build
is also Koji build and therefore it might be returned by `get_builds()`.