Allow inheriting the advisory ID from the dependent Freshmaker event
When submitting a manual rebuild, you can specify a dependent Freshmaker
event. If this is set, the `errata_id` value must match the advisory ID
associated with the dependent Freshmaker event. If it is not set, the
advisory ID is inherited.

As part of this, some additional input sanitization was added.
More input sanitization will come in a future commit.