Manager CLI doesn't depend on flask-script anymore
freshmaker/manage.py does not depend on flask-script from now on.
The original freshmaker-* commands are not changed, for example,

    freshmaker-upgradedb

Note that, FRESHMAKER_DEVELOPER_ENV have to be set to ensure the dev
config is loaded correctly for development.

During development, those manager commands can also run from manage.py
directly:

	python freshmaker/manage.py

Another change happens to command runssl. That is, app.run() is replaced
with the direct call to run_simple to avoid flask logs this warning and
stops the run.

    Silently ignoring app.run() because the application is run from the
    flask command line executable.

Signed-off-by: Chenxiong Qi <cqi@redhat.com>