Filter out images based on the content_set
Freshmaker queries Lightblue to get images containing affected RPMs
installed from a particular content_set. At the same time Freshmaker
asks to Lightblue also all the images with enabled the
auto_rebuild_tags tag (when not enabled the rebuilds of images in
this repository are disabled).
This gets done only because the Lightblue query will be easier and
cleaner this way. But because of that some images returned by that
query will not have the correct content_sets, for this reason we
need to filter out images based on the content_sets.

Signed-off-by: Giulia Naponiello <gnaponie@redhat.com>