Try hard to find the repositories for an image
An accurate list of repositories is important to ensure the
deduplication code works as expected.

If Freshmaker has built an image in the past, it may not have been
directly published and Lightblue won't list any repositories for it.
In such case, try to get a list of repositories from the original NVR.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>

WIP: adding intital unit tests

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>