Add the script to test new freshmaker deployment
Find out the latest successful event and get its the errata_id.
At the moment, when we release a new freshmaker version, we start a
dry-run build to manually check that the deployment was successful
curl --negotiate -u : -k -X POST -d '{"errata_id": 46932, "dry_run": true}'
https://freshmaker.dev.engineering.redhat.com/api/1/builds/ -l -v
But if the errata_id is old, it can happen that the build won't be
successful. This script is for this reason, it extract the errata_id
and then run that curl command. This way, we don't need to update the
documentation everytime this happens.