Make resolve_commit not set published property
A ContainerImage published property is resolved and set by
resolve_published. Therefore, if resolve_commit sets the published
property, there will be a side-effect when resolve_commit is called
again against an already resolved image, that is published property is
reset to None.

Whatever a ContainerImage is published or not, resolve_published will
always set the property. So, just simply remove ``published`` from
return value of _get_default_additional_data().

Resolves: FACTORY-6028

Signed-off-by: Chenxiong Qi <cqi@redhat.com>