Build container image without an RPM
Prior to this commit, the process of building a container image for
Freshmaker created a temporary RPM which was then installed in the
container image.

Given that the RPM is immediately discarded, and that Freshmaker is
never distributed as an RPM, this process is unnecessarily complex.

This commit removes the RPM build process completely and simply builds
the container image directly from source.

Because Freshmaker is not installed from an RPM, its contents are found
in a different location. Previously, the freshmaker-* console scripts
were installed under /usr/bin, but now they're under /usr/local/bin.
This is also true for the python package. It has moved from
/usr/lib/python3.7/site-packages/freshmaker to
/usr/local/lib/python3.7/site-packages/freshmaker

This change in path is expected as /usr/local is meant for content not
provided by the Linux distribution itself.

Also, man pages for the freshmaker-* console scripts are no longer
generated. There appears to be no value in generating those.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>

wip: will this work?!

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>