Periodically reduce size of images to rebuild
When list of lists with images to be rebuilt is constructed, we will
sometimes call deduplication function on it. Because we need to reduce
its size during construction, not after it.

For start I've chosen to do deduplication every 50 new tuples of images,
because deduplication doesn't reduce amount of tuples,
it only replaces images in this tuples with newer ones.

Signed-off-by: Andrei Paplauski <apaplaus@redhat.com>