Update parent image when parent image's version changes
The comment in _deduplicate_images_to_rebuild says:

    # 1) "handle_parent_change" - During this phase, we find out if update
    #    to latest image changes also the parent images.
    #    For example, foo-1-1 can be built against x-1-1, but foo-1-2 can
    #    be built against y-1-1. If we simply replace "foo-1-1" by "foo-1-2"
    #    while keeping the original parent image, the "foo-1-2" will be built
    #    against x-1-1 instead of y-1-1. This would be wrong.

the current implementation is updating the parent when parent is
changed from x-1-1 to y-1-1, where the name is changed, so why not
also update parent when only version changes?

With this change, if we have a case of:

    foo-1-1 was built against x-1-1
    foo-1-2 was built against x-2-1

then in additional to replace "foo-1-1" by "foo-1-2", we update parent
image to "x-2-1" as well.