Use Errata Tool CVE mapping functionality
Errata now provides an API to get builds by cve for an advisory, it
returns the builds affected by CVEs in advisory, we can use this API to
replace the old implementation of fetching packages affected by CVEs
and comparing the result with builds in advisory.

And we use the "security_impact" value from advisory instead of the
old "highest_cve_severity" which is fetched from sfm2. The
"security_impact" in errata is more precise, because for each CVE,
the severity we fetched from sfm2 is a global value, and for each CVE,
it can has different severity values for different products. So we
don't need the sfm2 module any more.

JIRA: CLOUDWF-2482