`get_latest_bundle` now uses `latest_in_channel` field
We now have the `latest_in_channel` field available which simplifies the
process of finding the bundles that are set as latest for some channel
in a index image.

During the making of this PR, it was discovered that the current way of
getting those bundles were incorrect. This commit fixes that. That's why
the changes does not exactly reflect the "version comparing" process
used before.

JIRA: CWFHEALTH-88