Check blocking advisories for overlapping images
Freshmaker while handling advisory with operator bundle will check not
only for original images, but also for images from blocking/previous
advisories and mark them for rebuild too. Because sometimes previously
rebuilt bundle image could have overriten original NVR of the
operator/operand images, so it might not be marked for rebuild event if
it should be.

JIRA: CWFHEALTH-60

Signed-off-by: Andrei Paplauski <apaplaus@redhat.com>