Override advisory state with state in state change message
When an ErrataAdvisoryStateChangedEvent is generated from an advisory
state change message, override the advisory state with new state in
message. This is because when there is message delay, a message of
advisory state change to other states (QE, REL_PREP, PUSH_READY) can
arrived after an advisory has already been shipped, and rebuilds can
be triggered for multiple times.

For example, there is an advisory 101, and its state was changed as
following:

1. 10:10:58: NEW_FILES -> QE, message A sent
2. 10:39:52: QE -> REL_PREP, message B sent
3. 10:45:07: REL_PREP -> PUSH_READY, message C sent
4. 10:46:22: PUSH_READY -> IN_PUSH, message D sent
5. 11:15:11: IN_PUSH -> SHIPPED_LIVE, message E sent

But there is something wrong on messaging, Freshmaker received the
first message A at 12:38:02, and it checked advisory state in Errata,
because at this moment, advisory 101 has been moved to "SHIPPED_LIVE",
message A will trigger image rebuilds, the other 4 messages received
after message A will also trigger rebuilds.