Get image digest from non-build registry repository
There may be a delay from the moment when a rebuilt operator/operand image
is shipped and it getting marked as published in Pyxis.

To workaround this, use the manifest list digest of a non-build registry
regardless of if the image is marked as published yet. This is safe since
manifest list digests will not change.