Allow manifest list digests from unpublished repositories
The issue is that there is a delay from when a BOTAS advisory
is shipped and when Pyxis is updated with the published
repositories entry on the image. The manifest list digest
is the same in both cases, so it's safe to fallback to the
unpublished manifest list digest value for this case.